Update README.md
Welcome to the Olympic Data Analyzer with Streamlit! This is an engaging and user-friendly web application designed to help you explore the rich history of the Summer and Winter Olympics. With this app, you can dive into Olympic data, visualize medal counts, and gain valuable insights into athlete statistics and the performance of countries and athletes over different Olympic seasons. Whether you're a sports enthusiast, data analyst, or just curious about Olympic history, this app offers an accessible and interactive way to unlock the stories hidden within the numbers of the world's most celebrated sporting events